refactor: move rope helpers into module
